# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AugmentedQuill",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/books/create": {
      "post": {
        "tags": ["Projects"],
        "summary": "Api Books Create",
        "description": "Handle the API request to books create.",
        "operationId": "api_books_create_api_v1_books_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookMutationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/books/delete": {
      "post": {
        "tags": ["Projects"],
        "summary": "Api Books Delete",
        "description": "Handle the API request to books delete.",
        "operationId": "api_books_delete_api_v1_books_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookMutationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/books/restore": {
      "post": {
        "tags": ["Projects"],
        "summary": "Api Books Restore",
        "description": "Handle the API request to books restore.",
        "operationId": "api_books_restore_api_v1_books_restore_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookRestoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookMutationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BookCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": ["name"],
        "title": "BookCreateRequest",
        "description": "Represents the BookCreateRequest type."
      },
      "BookDeleteRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": ["name"],
        "title": "BookDeleteRequest",
        "description": "Represents the BookDeleteRequest type."
      },
      "BookMutationResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "book_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Id"
          },
          "restore_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Restore Id"
          },
          "story": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StoryPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": ["ok"],
        "title": "BookMutationResponse",
        "description": "Response body for book create/delete/restore endpoints."
      },
      "BookRestoreRequest": {
        "properties": {
          "restore_id": {
            "type": "string",
            "title": "Restore Id"
          }
        },
        "type": "object",
        "required": ["restore_id"],
        "title": "BookRestoreRequest",
        "description": "Represents the BookRestoreRequest type."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Scene": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          },
          "beats": {
            "items": {
              "$ref": "#/components/schemas/SceneBeat"
            },
            "type": "array",
            "title": "Beats",
            "default": []
          },
          "active_characters": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Active Characters",
            "default": []
          },
          "passive_characters": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Passive Characters",
            "default": []
          },
          "sourcebook_entry_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sourcebook Entry Ids",
            "default": []
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time"
          },
          "scene_time": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SceneChronologyTime"
              },
              {
                "type": "null"
              }
            ]
          },
          "timeline_id": {
            "type": "string",
            "title": "Timeline Id",
            "default": "main"
          },
          "color_tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color Tag"
          },
          "prose_link": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SceneProseLink"
              },
              {
                "type": "null"
              }
            ]
          },
          "causes": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Causes",
            "default": []
          },
          "pinboard_x": {
            "type": "number",
            "title": "Pinboard X",
            "default": 100.0
          },
          "pinboard_y": {
            "type": "number",
            "title": "Pinboard Y",
            "default": 100.0
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "active"
          },
          "tag_personal_datetimes": {
            "items": {
              "$ref": "#/components/schemas/SceneTagPersonalDatetime"
            },
            "type": "array",
            "title": "Tag Personal Datetimes"
          },
          "annotations": {
            "items": {
              "$ref": "#/components/schemas/SceneAnnotation"
            },
            "type": "array",
            "title": "Annotations",
            "description": "Optional inline annotations associated with this scene."
          }
        },
        "type": "object",
        "required": ["id"],
        "title": "Scene",
        "description": "A narrative scene used for structural story planning.\n\nActive characters, passive characters, location, and time are stored as\nsourcebook entry IDs so the frontend can look them up by reference.\n\n``causes`` stores IDs of other scenes that this scene causally precedes.\nThese are not narrative order indexes; they are dependency constraints that\nallow the pinboard to highlight cause/effect relationships.\n\n``pinboard_x`` / ``pinboard_y`` store the card's free-form position on the\npinboard canvas, in logical (unscaled) units."
      },
...
      "SceneBeat": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Stable beat identifier. Usually generated automatically."
          },
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Short description of the beat's action or event."
          },
          "prose_link": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SceneProseLink"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional link from this beat to a specific prose range."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["text"],
        "title": "SceneBeat",
        "description": "A single beat within a scene \u2013 a discrete micro-action or plot step."
      },
...
      "SceneChronologyTime": {
        "properties": {
          "temporal_zoned_datetime": {
            "type": "string",
            "title": "Temporal Zoned Datetime",
            "description": "Normalized ISO 8601 timestamp for the scene's chronology."
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value",
            "description": "Optional shorthand scene time string accepted by the tool. When provided, it is normalized into temporal_zoned_datetime."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["temporal_zoned_datetime"],
        "title": "SceneChronologyTime",
        "description": "Scene-local timeline point represented as a Temporal ZonedDateTime string."
      },
...
      "StoryBook": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "folder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folder"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "chapters": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/StoryChapterSummary"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapters"
          }
        },
        "type": "object",
        "title": "StoryBook",
        "description": "Book descriptor embedded in the story payload."
      },
      "StoryChapterSummary": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "book_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Id"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "private_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Private Notes"
          },
          "conflicts": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conflicts"
          }
        },
        "type": "object",
        "title": "StoryChapterSummary",
        "description": "Lightweight chapter descriptor embedded in the story payload."
      },
...
      "StoryLLMPrefs": {
        "properties": {
          "prompt_overrides": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Overrides"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "max_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Tokens"
          }
        },
        "type": "object",
        "title": "StoryLLMPrefs",
        "description": "Per-project LLM preference overrides embedded in story data."
      },
      "StoryPayload": {
        "properties": {
          "project_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Title"
          },
          "story_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Story Summary"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "private_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Private Notes"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "annotations": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Annotations"
          },
          "image_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Style"
          },
          "image_additional_info": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Additional Info"
          },
          "project_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Type"
          },
          "books": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/StoryBook"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Books"
          },
          "sourcebook": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/StorySourcebookEntry"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcebook"
          },
          "conflicts": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conflicts"
          },
          "llm_prefs": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StoryLLMPrefs"
              },
              {
                "type": "null"
              }
            ]
          },
          "chapters": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/StoryChapterSummary"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapters"
          },
          "scenes": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Scene"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenes"
          }
        },
        "type": "object",
        "title": "StoryPayload",
        "description": "Story data as returned by project selection and mutation endpoints.\n\nThis matches the shape produced by\n``services.projects.project_helpers.normalize_story_for_frontend``.\nAll fields are optional because the payload depends on what the author\nhas filled in for a given project."
      },
      "StorySourcebookEntry": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "synonyms": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Synonyms"
          },
          "images": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Images"
          },
          "keywords": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keywords"
          },
          "relations": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relations"
          },
          "origin_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Date"
          },
          "destination_datetime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Datetime"
          },
          "destination_relative": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Relative"
          },
          "creates_new_timeline": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Creates New Timeline"
          },
          "timeline_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeline Id"
          }
        },
        "type": "object",
        "title": "StorySourcebookEntry",
        "description": "Minimal sourcebook entry embedded in the story payload."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}